+++ python/pyxel/editor/image_editor.py
@@ -128,9 +128,12 @@ class ImageEditor(EditorBase):
         colors = list(pyxel.colors)
         user_colors = colors[pyxel.NUM_COLORS :]
         pyxel.colors[:] = user_colors
-        pyxel.images[self.image_index_var].load(
-            self.focus_x_var * 8, self.focus_y_var * 8, filename
-        )
+        try:
+            pyxel.images[self.image_index_var].load(
+                self.focus_x_var * 8, self.focus_y_var * 8, filename
+            )
+        except Exception:
+            pass
         pyxel.colors[:] = colors
 
     def __on_update(self):