+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 2.8.2
 
 - Prioritized system SDL2 over bundled SDL2 on Linux
+- Added external file drop support for web-based editor
+- Fixed crash when dropping non-image files on the image editor
 
 ## 2.8.1